Solved problem of shit elements on the 3D mesh. Now all the calculations, that includes the normal to the surface and the kappa, all the structures and vectores, are done using only the surface mesh. The code is working without stop due to segmentation fault and divergence of linear system. On Simulator3D we need to set for bubble rising the relation of:
dt*invMLumped*fint + dt*invMrhoLumped*gravity

for smooth velocity and pressure profile since we are assembling the K and M
matriz with mu and rho constant to respect of each phase, then the interface
is well and sharpy defined. The relation:

invA*fint + invA*gravity

works for droplet in 2D case. The 3D is not tested yet.
Bogdan's test case (thesis page 79) is benchmark in mainBubble giving
great results.