This commit is basically a bug fix. The work has been performed in the commit:
40cd26876931d32a74563baa57fe8e937bdb8c95

And the code was adjusted to run with no problems. All the scrips were
tested and seem to be working properly. The bug was related to the
Model3D class and its methods:

- curvature calculation when new point is added. The code now is taking
  the average curvature between two vertices and setting to the new one.
- the method delete3DPoints was not set and all the deleted 3D mesh
  points were not deleted even performing removePointByVolume,
  ByDiffusion etc.
- the method removePointByVolume is not being used anymore because it
  damages the TETGEN mesh after several time steps.

Also the mesh treatment (insertion,deletion,flipping and contraction)
for both structures (surface and 3D) are now done inside the script
files (Ex. mainBubble.cpp). Then it gives more flexibility to individual
cases to set the appropriated case.