In TElement is now possible to change the element type by changing the number of NUMGLEU from 5 (mini element) to 10 (quad element). TElement is spread to all the required classes to adapt from the define (pre processor diretive) and run the element case.
Model3D: All the simple BC (except those that are using boundarVert) are
configured to work with two elements: mini and quad.

SemiLagrangian: Configured to work with two elements by defining it into
TElement.

Simulator3D: many changes and bug fix. The most important is the
settings of DT that should be done before move the grid points. That's
why the moving point function was taken out from all the stepALE and
stepLagrangian methods and then create movePoints() method. All the main
script files are updated to work of such a modification.